refactor(daemon): split lib.rs into log_init / startup / shutdown sibling modules
`crates/uffs-daemon/src/lib.rs` was 1066 LOC (266 over the 800-LOC
file-size policy ceiling).  The existing rationale comment argued
against splitting the `spawn_*` cluster because doing so would
fragment the parent-task lifetime relationships — and that argument
still holds.  But the orchestrator's *setup* and *terminal* phases
have no such coupling and can move out cleanly.

## Extractions

* `crates/uffs-daemon/src/log_init.rs` (105 LOC, new) — `init_tracing`
  + `default_log_file`.  No collision with the `tracing` crate
  because the module is named `log_init`.  `pub use
  log_init::init_tracing;` re-export preserves the public API
  exactly (`uffs_daemon::init_tracing` still works for the daemon
  binary + the embedded `uffs daemon run` subcommand).

* `crates/uffs-daemon/src/startup.rs` (238 LOC, new) — pre-spawn
  helpers:

    - `install_catastrophe_panic_hook`
    - `log_daemon_starting` + `emit_daemon_starting_event`
    - `load_daemon_config`
    - `bootstrap_lifecycle_manager`
    - `gather_mft_files` + `drive_letter_matches` + `resolve_drive_list`
      (windows + non-windows variants)
    - `validate_data_sources`

  All `pub(crate)`.  `drive_letter_matches` is `pub(crate)` (was
  private) because the existing regression-pin test in
  `crates/uffs-daemon/src/tests.rs` exercises it directly — the
  contract pin was preserved by updating the import to
  `super::startup::drive_letter_matches;`.

* `crates/uffs-daemon/src/shutdown.rs` (84 LOC, new) —
  `await_shutdown_then_force_exit` + `force_exit_with_watchdog`.
  All `pub(crate)`.

## lib.rs after

* **689 LOC** (was 1066) — under the 800-LOC ceiling without an
  exception.
* Keeps the `spawn_*` cluster cohesive:
  `spawn_load_task` / `spawn_ipc_servers` /
  `spawn_stats_heartbeat` / `spawn_idle_demote_controller` /
  `spawn_journal_loops_for_warm_shards` /
  `spawn_pressure_subscriber`.
* Keeps `DaemonConfig` (public struct) and `run_daemon` (public
  orchestrator) for binary + embedded use.
* Removed the obsolete `Exception: file_size_policy allows this
  file to exceed 800 LOC` rationale comment from the crate-root
  rustdoc.

## file_size_exceptions.txt

Removed the now-obsolete `crates/uffs-daemon/src/lib.rs|PERMANENT:
…` entry.  Local `bash scripts/ci/check_file_size_policy.sh` passes
clean.

## Rule-1 adherence

* Zero `#[allow(...)]` introductions.
* No suppression hacks, no skipped tests.
* No public API change — `uffs_daemon::{init_tracing, run_daemon,
  DaemonConfig}` resolve to the same items as before via the
  `pub use log_init::init_tracing;` re-export.
* `cargo fmt --all` — clean.
* `cargo clippy --workspace --all-targets -- -D warnings` — clean.
* `cargo test -p uffs-daemon --lib` — **298 passed / 0 failed**.
* `bash scripts/ci/check_file_size_policy.sh` — clean.

Refs #302 (Phase 10 umbrella).
